Finshed Timezone settings & some cleanup

+++ src/main/java/sh/niall/misty/utils/ui/Helper.java
@@ -5,13 +5,13 @@ import java.util.Random;
 
 public class Helper {
 
-    private static Color[] colors = new Color[] {
-      Color.BLUE,
-      Color.CYAN,
-      Color.GREEN,
-      Color.MAGENTA,
-      Color.ORANGE,
-      Color.PINK,
+    private static Color[] colors = new Color[]{
+            Color.BLUE,
+            Color.CYAN,
+            Color.GREEN,
+            Color.MAGENTA,
+            Color.ORANGE,
+            Color.PINK,
     };
     private static Random random = new Random();
 
@@ -23,7 +23,7 @@ public class Helper {
      * @param plural The plural word to use
      * @return The correct word to use based on the count. Returns single if count == 1
      */
-    public static String singularPlural(int count, String single, String plural) {
+    public static String singularPlural(long count, String single, String plural) {
         if (count == 1)
             return single;
         return plural;
